Move mesh generation to GPU

+++ Lectures/Lecture3/Assets/Scripts/MarchingCubes.Tables.cs
@@ -49,7 +49,7 @@ public static class MarchingCubes
         /// <summary>
         /// How many triangles should be there in a cube for a given CaseIndex?
         /// </summary>
-        public static readonly byte[] CaseToTrianglesCount = {
+        public static readonly int[] CaseToTrianglesCount = {
             0, 1, 1, 2, 1, 2, 2, 3, 1, 2, 2, 3, 2, 3, 3, 2, 1, 2, 2, 3, 2, 3, 3, 4, 2, 3, 3, 4, 3, 4, 4, 3,
             1, 2, 2, 3, 2, 3, 3, 4, 2, 3, 3, 4, 3, 4, 4, 3, 2, 3, 3, 2, 3, 4, 4, 3, 3, 4, 4, 3, 4, 5, 5, 2,
             1, 2, 2, 3, 2, 3, 3, 4, 2, 3, 3, 4, 3, 4, 4, 3, 2, 3, 3, 4, 3, 4, 4, 5, 3, 4, 4, 5, 4, 5, 5, 4,
